+++ core/src/main/java/iwf/core/WorkflowStartOptions.java
@@ -4,5 +4,5 @@ import org.immutables.value.Value;
 
 @Value.Immutable
 public interface WorkflowStartOptions {
-    Long getWorkflowTimeoutSeconds();
+    Integer getWorkflowTimeoutSeconds();
 }
